Add step (http) timeout
The step timeout is effectively passed to every step's eval() method, but only the http step  uses it.  By default its passed in from the testing service similar to the test timeout, but unlike the test timeout, it can be overridden on individual http requests via the options object inside the http step.