Add retry feature when reporting test results
If the POST request to the testing service fails, wait 3 seconds and retry. Continue to retry up to 10 times before giving up.  This should help reduce the number of tests that fail after they have started.

If the POST request to the testing service fails, wait 3 seconds and retry. Continue to retry up to 10 times before giving up.  This should help reduce the number of tests that fail after they have started.