Add safeParameters object as alternative
One of the downsides to just filtering using removeSensitiveData is that it can not distinguish between non-sensitive strings and sensitive string. If a user had a short password then it might replace non-sensitive substrings as well. This would allow easy guessing of very poor passwords.

`this.safeParameters` is a structured object with the passwords removed which is an easy swap in the actions for this.parameters where short/long description is used